requirements.txt
# Web framework
fastapi==0.115.12
uvicorn==0.34.2
jinja2==3.1.6
python-multipart==0.0.20

# News collection
feedparser==6.0.11
trafilatura==2.0.0
duckduckgo-search==7.5.3

# NotebookLM
notebooklm-py[browser]==0.3.4

# Google Drive
google-api-python-client==2.170.0
google-auth-oauthlib==1.2.1

# Scheduler
apscheduler==3.11.0

# Environment
python-dotenv==1.1.0

# Database
aiosqlite==0.21.0
